refactor(multi): sweep 500-1000L god files in clawde/cli/plugins-pro (T-E2-06)
Split hardening_check.go (808L) → hardening_check_db.go /
hardening_check_auth_net.go / hardening_check_infra.go /
hardening_check_helpers.go in cli/internal/doctor.

Split loader.go (942L) → loader_known_vars.go / loader_parse_env.go /
loader_helpers.go in cli/internal/config.

Split scaffold.go (938L) → scaffold_templates_infra.go /
scaffold_templates_code.go in cli/internal/plugin/scaffold.

Each file has a Purpose/Inputs/Outputs/Constraints/SPORT comment block.
Build verified clean (go build ./...).
