+++ internal/plugin/scaffold/scaffold.go
@@ -5,6 +5,16 @@
 // Both entry points call scaffold.Run with a Params struct; the output is
 // identical so that plugin authors get the same result regardless of whether
 // they have nself installed or use the SDK devkit directly.
+//
+// Purpose: Core types and all logic functions for plugin scaffold generation.
+//          Template strings are in scaffold_templates_infra.go (infrastructure,
+//          devops, metadata templates) and scaffold_templates_code.go (Go code
+//          templates: main, config, server, server_test).
+// Inputs:  Options struct — name, tier, language, tenancy mode, overrides.
+// Outputs: Result struct — output directory path and list of emitted files.
+// Constraints: Must remain import-compatible with plugin-sdk-go/devkit.
+//              Template strings must live in the _templates_*.go files, not here.
+// SPORT:   cli/internal/plugin/scaffold — decomposed from scaffold.go (T-E2-06).
 package scaffold
 
 import (
@@ -480,459 +490,3 @@ func licenseForTier(tier string) string {
 	}
 	return "MIT"
 }
-
-// --- tenancy + plugin.json templates ---
-
-// tmplPluginJSON is the plugin.json template. It requires a struct with all
-// Params fields plus MultiAppSupported (bool) and IsolationColumn (string).
-const tmplPluginJSON = `{
-  "name": "{{.Name}}",
-  "version": "0.1.0",
-  "description": "{{.Description}}",
-  "author": "{{.Author}}",
-  "license": {{if eq .Tier "pro"}}"Source-Available"{{else}}"MIT"{{end}},
-  "isCommercial": {{if eq .Tier "pro"}}true{{else}}false{{end}},
-  {{- if eq .Tier "pro"}}
-  "licenseType": "pro",
-  "requiredEntitlements": ["pro"],
-  "requires_license": true,
-  {{- end}}
-  "homepage": "https://nself.org/plugins",
-  "minNselfVersion": "{{.MinCLI}}",
-  "minSdkVersion": "{{.MinSDK}}",
-  "category": "{{.Category}}",
-  {{- if .Bundle}}
-  "bundle": "{{.Bundle}}",
-  {{- end}}
-  "multiApp": {
-    "supported": {{.MultiAppSupported}},
-    "isolationColumn": "{{.IsolationColumn}}"
-  },
-  "tags": ["{{.Name}}"]
-}
-`
-
-// tmplMigrationNone is emitted when the plugin stores no per-user Postgres data.
-const tmplMigrationNone = `-- {{.Name}} initial migration
--- No multi-tenancy columns required for this plugin.
--- Add your schema here.
-
-CREATE TABLE IF NOT EXISTS np_{{.Name}}_items (
-    id          UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    created_at  TIMESTAMPTZ NOT NULL DEFAULT now(),
-    updated_at  TIMESTAMPTZ NOT NULL DEFAULT now()
-);
-`
-
-// tmplMigrationAppIsolation is emitted for multi-app isolation within one nSelf deploy.
-// Uses source_account_id per the Multi-Tenant Convention Wall.
-const tmplMigrationAppIsolation = `-- {{.Name}} initial migration
--- Multi-app isolation: source_account_id separates independent consumer apps
--- within one nSelf deploy. See: docs/architecture/multi-tenant-conventions.md
-
-CREATE TABLE IF NOT EXISTS np_{{.Name}}_items (
-    id                 UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    source_account_id  TEXT NOT NULL DEFAULT 'primary',
-    created_at         TIMESTAMPTZ NOT NULL DEFAULT now(),
-    updated_at         TIMESTAMPTZ NOT NULL DEFAULT now()
-);
-
--- Row-level security: each app sees only its own rows.
-ALTER TABLE np_{{.Name}}_items ENABLE ROW LEVEL SECURITY;
-
-CREATE POLICY np_{{.Name}}_items_isolation ON np_{{.Name}}_items
-    USING (source_account_id = current_setting('app.source_account_id', true));
-`
-
-// tmplMigrationCloudTenant is emitted for Cloud SaaS tenancy.
-// Uses tenant_id UUID per the Multi-Tenant Convention Wall.
-const tmplMigrationCloudTenant = `-- {{.Name}} initial migration
--- Cloud multi-tenancy: tenant_id separates paying customers in nSelf Cloud SaaS.
--- See: docs/architecture/multi-tenant-conventions.md
--- NEVER use tenant_id for per-app isolation within one deploy — use the app-isolation column instead.
-
-CREATE TABLE IF NOT EXISTS np_{{.Name}}_items (
-    id          UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    tenant_id   UUID,
-    created_at  TIMESTAMPTZ NOT NULL DEFAULT now(),
-    updated_at  TIMESTAMPTZ NOT NULL DEFAULT now()
-);
-
--- Row-level security: each tenant sees only its own rows.
-ALTER TABLE np_{{.Name}}_items ENABLE ROW LEVEL SECURITY;
-
-CREATE POLICY np_{{.Name}}_items_tenant ON np_{{.Name}}_items
-    USING (tenant_id = current_setting('app.tenant_id', true)::UUID);
-`
-
-// tmplMigrationBoth emits both columns for developers who are unsure which
-// convention they need. Remove the unused column before going to production.
-const tmplMigrationBoth = `-- {{.Name}} initial migration
--- Both multi-tenancy columns included. Remove the one you do not need before
--- going to production. See: docs/architecture/multi-tenant-conventions.md
---
---   source_account_id: per-app isolation within one nSelf deploy
---   tenant_id:         Cloud SaaS paying-customer isolation
-
-CREATE TABLE IF NOT EXISTS np_{{.Name}}_items (
-    id                 UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    source_account_id  TEXT NOT NULL DEFAULT 'primary',
-    tenant_id          UUID,
-    created_at         TIMESTAMPTZ NOT NULL DEFAULT now(),
-    updated_at         TIMESTAMPTZ NOT NULL DEFAULT now()
-);
-
-ALTER TABLE np_{{.Name}}_items ENABLE ROW LEVEL SECURITY;
-
--- Choose ONE of the two policies below and delete the other.
-CREATE POLICY np_{{.Name}}_items_isolation ON np_{{.Name}}_items
-    USING (source_account_id = current_setting('app.source_account_id', true));
-
--- CREATE POLICY np_{{.Name}}_items_tenant ON np_{{.Name}}_items
---     USING (tenant_id = current_setting('app.tenant_id', true)::UUID);
-`
-
-// tmplHasuraNoFilter is the Hasura metadata stub for plugins that use
-// source_account_id (app isolation) or no tenancy at all. No tenant row filter
-// is required because isolation is handled at the Postgres RLS layer.
-const tmplHasuraNoFilter = `{
-  "table": {
-    "schema": "public",
-    "name": "np_{{.Name}}_items"
-  },
-  "select_permissions": [
-    {
-      "role": "user",
-      "permission": {
-        "columns": "*",
-        "filter": {}
-      }
-    }
-  ]
-}
-`
-
-// tmplHasuraCloudFilter is the Hasura metadata stub for Cloud multi-tenant
-// plugins. The row filter enforces that each tenant only sees its own rows via
-// the X-Hasura-Tenant-Id session variable.
-const tmplHasuraCloudFilter = `{
-  "table": {
-    "schema": "public",
-    "name": "np_{{.Name}}_items"
-  },
-  "select_permissions": [
-    {
-      "role": "user",
-      "permission": {
-        "columns": "*",
-        "filter": {
-          "tenant_id": {
-            "_eq": "X-Hasura-Tenant-Id"
-          }
-        }
-      }
-    }
-  ]
-}
-`
-
-// --- template strings ---
-
-const tmplCompose = `# docker-compose.plugin.yml for {{.Name}}
-# Merged into the generated stack by ` + "`nself build`" + `. Do not hand-edit.
-services:
-  {{.Name}}:
-    image: nself/{{.Name}}:${{"{"}}{{.EnvPrefix}}_VERSION:-latest}
-    container_name: ${PROJECT_NAME:-nself}_{{.Name}}
-    restart: unless-stopped
-    environment:
-      LOG_LEVEL: ${LOG_LEVEL:-info}
-      DATABASE_URL: ${DATABASE_URL}
-      {{.EnvPrefix}}_LISTEN_ADDR: ":{{.Port}}"
-    ports:
-      - "127.0.0.1:{{.Port}}:{{.Port}}"
-    networks:
-      - nself_net
-networks:
-  nself_net:
-    external: true
-`
-
-const tmplDockerignore = `.git
-.gitignore
-README.md
-Dockerfile
-docker-compose*.yml
-.air.toml
-tmp/
-*.test
-coverage.out
-`
-
-const tmplAirToml = `# air.toml — hot-reload for {{.Name}} dev (pair with nself plugin dev)
-root = "."
-tmp_dir = "tmp"
-
-[build]
-  cmd = "go build -o ./tmp/{{.Name}} ./cmd"
-  bin = "tmp/{{.Name}}"
-  delay = 500
-  include_ext = ["go", "yaml", "yml"]
-  exclude_dir = ["tmp", "vendor", ".git"]
-
-[log]
-  time = true
-
-[color]
-  app = "magenta"
-`
-
-const tmplReadme = `# {{.PascalName}} Plugin
-
-{{.Description}}
-
-Tier: ` + "`{{.Tier}}`" + `{{if .Bundle}}  ·  Bundle: ` + "`{{.Bundle}}`" + `{{end}}  ·  Category: ` + "`{{.Category}}`" + `
-
-## Local development
-
-` + "```bash" + `
-go mod tidy
-go test ./...
-go run ./cmd        # runs on :{{.Port}}
-` + "```" + `
-
-With hot-reload (install [air](https://github.com/air-verse/air)):
-
-` + "```bash" + `
-nself plugin dev {{.Name}}
-` + "```" + `
-
-## Endpoints
-
-- ` + "`GET /healthz`" + ` — liveness
-- ` + "`GET /readyz`" + ` — readiness
-- ` + "`GET /metrics`" + ` — Prometheus metrics
-- ` + "`GET /version`" + ` — plugin version
-- ` + "`GET /v1/hello`" + ` — starter handler
-
-## License
-
-{{if eq .Tier "pro"}}Source-Available (pro tier). Requires an active nSelf license key.{{else}}MIT.{{end}}
-`
-
-const tmplCI = `name: CI
-on: [push, pull_request]
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-go@v5
-        with:
-          go-version: "1.23"
-      - name: Test
-        run: go test ./...
-      - name: Build Docker image
-        run: docker build -t {{.Name}}:test .
-      - name: Health check
-        run: |
-          docker run -d -p {{.Port}}:{{.Port}} --name test {{.Name}}:test
-          sleep 2
-          curl -f http://localhost:{{.Port}}/healthz
-          docker stop test
-`
-
-const tmplMain = `// Package main is the entrypoint for the {{.Name}} plugin.
-package main
-
-import (
-	"context"
-	"errors"
-	"net/http"
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
-
-	"github.com/nself-org/{{.RepoBucket}}/{{.Name}}/internal/config"
-	"github.com/nself-org/{{.RepoBucket}}/{{.Name}}/internal/server"
-
-	"github.com/nself-org/plugin-sdk-go/logger"
-)
-
-// Version is stamped at build time via -ldflags.
-var Version = "0.1.0"
-
-func main() {
-	cfg := config.FromEnv()
-	log := logger.New(logger.Options{
-		Plugin:  "{{.Name}}",
-		Version: Version,
-		Level:   logger.ParseLevel(cfg.LogLevel),
-	})
-
-	srv := server.New(server.Deps{Config: cfg, Logger: log, Version: Version})
-
-	httpSrv := &http.Server{
-		Addr:              cfg.ListenAddr,
-		Handler:           srv,
-		ReadHeaderTimeout: 10 * time.Second,
-	}
-
-	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
-	defer stop()
-
-	errCh := make(chan error, 1)
-	go func() {
-		defer func() {
-			if r := recover(); r != nil {
-				errCh <- fmt.Errorf("server goroutine panic: %v", r)
-			}
-		}()
-		log.Info("{{.Name}} listening", "addr", cfg.ListenAddr)
-		errCh <- httpSrv.ListenAndServe()
-	}()
-
-	select {
-	case <-ctx.Done():
-		log.Info("shutdown signal received")
-	case err := <-errCh:
-		if err != nil && !errors.Is(err, http.ErrServerClosed) {
-			log.Error("server failed", "error", err)
-		}
-	}
-
-	shutdownCtx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
-	if err := httpSrv.Shutdown(shutdownCtx); err != nil {
-		log.Error("graceful shutdown failed", "error", err)
-		os.Exit(1)
-	}
-	log.Info("{{.Name}} stopped cleanly")
-}
-`
-
-const tmplConfig = `// Package config loads {{.Name}} config from environment variables.
-package config
-
-import (
-	"fmt"
-	"os"
-	"strings"
-)
-
-// Config holds runtime config.
-type Config struct {
-	ListenAddr  string
-	LogLevel    string
-	DatabaseURL string
-}
-
-// FromEnv reads config from env vars with sensible defaults.
-func FromEnv() Config {
-	return Config{
-		ListenAddr:  envOr("{{.EnvPrefix}}_LISTEN_ADDR", ":{{.Port}}"),
-		LogLevel:    envOr("LOG_LEVEL", "info"),
-		DatabaseURL: os.Getenv("DATABASE_URL"),
-	}
-}
-
-// Validate returns an error if required fields are missing.
-func (c Config) Validate() error {
-	if c.ListenAddr == "" {
-		return fmt.Errorf("{{.Name}}: listen address must not be empty")
-	}
-	return nil
-}
-
-func envOr(key, def string) string {
-	if v := strings.TrimSpace(os.Getenv(key)); v != "" {
-		return v
-	}
-	return def
-}
-`
-
-const tmplServer = `// Package server wires the HTTP router for the {{.Name}} plugin.
-package server
-
-import (
-	"context"
-	"log/slog"
-	"net/http"
-
-	"github.com/go-chi/chi/v5"
-
-	"github.com/nself-org/{{.RepoBucket}}/{{.Name}}/internal/config"
-
-	sdkmetrics "github.com/nself-org/plugin-sdk-go/metrics"
-	sdkserver "github.com/nself-org/plugin-sdk-go/server"
-)
-
-// Deps wires runtime dependencies.
-type Deps struct {
-	Config  config.Config
-	Logger  *slog.Logger
-	Version string
-}
-
-type readyFn func(ctx context.Context) error
-
-func (f readyFn) Ready(ctx context.Context) error { return f(ctx) }
-
-// New returns a ready-to-serve http.Handler.
-func New(d Deps) http.Handler {
-	return sdkserver.New(sdkserver.Options{
-		Plugin:  "{{.Name}}",
-		Version: d.Version,
-		Ready: readyFn(func(ctx context.Context) error {
-			return d.Config.Validate()
-		}),
-		Routes: func(r chi.Router, m *sdkmetrics.Registry) {
-			r.Route("/v1", func(r chi.Router) {
-				r.With(m.Middleware("/v1/hello")).Get("/hello", helloHandler(d.Logger))
-			})
-		},
-	})
-}
-
-func helloHandler(log *slog.Logger) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if log != nil {
-			log.Info("hello called", "method", r.Method, "path", r.URL.Path)
-		}
-		w.Header().Set("Content-Type", "application/json")
-		_, _ = w.Write([]byte(` + "`" + `{"plugin":"{{.Name}}","hello":"world"}` + "`" + `))
-	}
-}
-`
-
-const tmplServerTest = `package server
-
-import (
-	"net/http/httptest"
-	"testing"
-
-	"github.com/nself-org/{{.RepoBucket}}/{{.Name}}/internal/config"
-)
-
-func TestHelloEndpoint(t *testing.T) {
-	h := New(Deps{Config: config.Config{ListenAddr: ":{{.Port}}"}, Version: "test"})
-	req := httptest.NewRequest("GET", "/v1/hello", nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	if rr.Code != 200 {
-		t.Fatalf("expected 200, got %d", rr.Code)
-	}
-}
-
-func TestHealthz(t *testing.T) {
-	h := New(Deps{Config: config.Config{ListenAddr: ":{{.Port}}"}, Version: "test"})
-	req := httptest.NewRequest("GET", "/healthz", nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	if rr.Code != 200 {
-		t.Fatalf("expected 200 from /healthz, got %d", rr.Code)
-	}
-}
-`
